Move to a helper instead a component for waves
