--- app/components/waves-effect.hbs
@@ -1,1 +1,0 @@
-{{yield}}
