--- app/components/waves-effect.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-bootstrap-components/components/waves-effect';
