--- tests/integration/components/waves-effect-test.js
@@ -1,27 +1,0 @@
-import { moduleForComponent, test } from 'ember-qunit';
-import hbs from 'htmlbars-inline-precompile';
-
-
-moduleForComponent('waves-effect', 'Integration | Component | waves effect', {
-  integration: true
-});
-
-test('it renders', function(assert) {
-  assert.expect(2);
-
-  // Set any properties with this.set('myProperty', 'value');
-  // Handle any actions with this.on('myAction', function(val) { ... });
-
-  this.render(hbs`{{waves-effect}}`);
-
-  assert.equal(this.$().text(), '');
-
-  // Template block usage:
-  this.render(hbs`
-    {{#waves-effect}}
-      template block text
-    {{/waves-effect}}
-  `);
-
-  assert.equal(this.$().text().trim(), 'template block text');
-});
